+++ KNXEngine/protocol/KNXUtils.js
@@ -13,17 +13,28 @@ const splitIP = (ip, name = 'ip') => {
 };
 exports.splitIP = splitIP;
 const validateKNXAddress = (address, isGroup = false) => {
+    // 22/12/2021 Supergiovane: https://support.knx.org/hc/en-us/articles/115003188109-Group-Addresses
     if (typeof (address) === 'string') {
         const digits = address.split(/[./]/);
-        if (digits.length < 2 || digits.length > 3) {
-            throw new Error(`Invalid address format: ${address}`);
+        if (digits.length !== 3) {
+            throw new Error(`Invalid address format: ${address} Only 3 level addresses are allowed`);
         }
         let count = 0;
         let newAddress = 0;
         for (let i = digits.length - 1; i >= 0; i--, count++) {
             const digit = Number(digits[i]);
-            if (isNaN(digit) || (count > 1 && digit > 15) || (count === 0 && digit > 255)) {
-                throw new Error(`Invalid digit at pos ${i} inside address: ${address}`);
+            if (isGroup) {
+                // Validating Group Address
+                if (isNaN(digit) || (count === 2 && digit > 31) || (count === 1 && digit > 7) || (count === 0 && digit > 255)) {
+                    // 22/12/2021 Supergiovane disabled digits validation
+                    throw new Error(`Invalid GA digit ${digit} inside address: ${address}`);
+                }
+            } else {
+                // Validating KNX Device Address
+                if (isNaN(digit) || (count > 1 && digit > 15) || (count === 0 && digit > 255)) {
+                    // 22/12/2021 Supergiovane disabled digits validation
+                    throw new Error(`Invalid Individual Address digit ${digit} inside address: ${address}`);
+                }
             }
             if (count === 0) {
                 newAddress = digit;